python/requirements.txt
attrs==22.2.0
exceptiongroup==1.1.1
iniconfig==2.0.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.2
tomli==2.0.1
